+++ signal_windows.go
@@ -1,0 +1,4 @@
+package main
+
+// not support windows
+func signalHandler() {}
